### 2018-09-11 #936 @whedon
[PRE REVIEW]: ungroup: An R package for efficient estimation of smooth distributions from coarsely binned data
**Submitting author:** @mpascariu (<a href="http://orcid.org/0000-0002-2568-6489">Marius Dan Pascariu</a>)
**Repository:** <a href="https://github.com/mpascariu/ungroup" target ="_blank">https://github.com/mpascariu/ungroup</a>
**Version:** v1.0.3
