### 2018-09-11 #936 @arfon
@mpascariu - please update your YAML paper header to exactly match this structure (looks like the `authors` field is missing) https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
