### 2018-09-11 c9d704a
update YAML paper header to exactly match joss structure
