### 2018-09-11 #936 @mpascariu
Seems to be an issue with the `tiff` files. Should I try saving the figures as `pdf` files?
