+++ joss/paper.md
@@ -35,11 +35,11 @@ bibliography: paper.bib
 
 Ungrouping binned data can be desirable for many reasons: Bins can be too coarse to allow for accurate analysis; comparisons can be hindered when different grouping approaches are used in different histograms; the last interval may be wide and open-ended masking the tail behaviour of the underlying distribution. Age-at-death distributions grouped into age classes and abridged life tables are examples of binned data which can be ungrouped with the package ``ungroup``. The modest assumptions of the methodology underpinning the PCLM method make it suitable for many demographic and epidemiological applications. For a detailed description of the method and applications see @rizzi2015 and @rizzi2016.
 
-![Ungrouping of the age-at-death distribution and estimating age-specific death rates. The original death counts and exposures taken from the @hmd2018 using the ``MortalityLaws`` R package [-@MortalityLaws160] were grouped in 5-year bins plus a wide class for ages 85+. In each panel, the original aggregated data is compared with smoothly estimated values.](pclm1D.tiff)
+![Ungrouping of the age-at-death distribution and estimating age-specific death rates. The original death counts and exposures taken from the @hmd2018 using the ``MortalityLaws`` R package [-@MortalityLaws160] were grouped in 5-year bins plus a wide class for ages 85+. In each panel, the original aggregated data is compared with smoothly estimated values.](figures/pclm1D.pdf)
 
 The penalized composite link model can be extended to a two-dimensional regression problem [@rizzi2018]. The two-dimensional regression analysis combines two approaches: the PCLM for ungrouping in one dimension and two-dimensional smoothing with P-splines [@currie2004]. As an example one can ungroup age-specific distributions from the coarsely grouped data and smooth across adjacent calendar years to estimate both detailed age-at-death distributions and mortality time trends.
 
-![Two-dimensional ungrouping of the age-at-death distributions and mortality surface. The 3-D figures are generate using the ``rgl`` R package [-@rgl09916].](pclm2D.tiff)
+![Two-dimensional ungrouping of the age-at-death distributions and mortality surface. The 3-D figures are generate using the ``rgl`` R package [-@rgl09916].](figures/pclm2D.pdf)
 
 
 # Acknowledgment
