### 2018-09-11 #936 @mpascariu
I have generated figures in pdf format. Hope it will work now.
