### 2018-09-11 #936 @arfon
@mpascariu - layout is tricky with Markdown/Pandoc. I suggest we wait until the review is complete before tweaking layout.
