### 2018-09-11 #936 @karthik
@mpascariu I'm looking for reviewers now so stay tuned. Here are some pre-review suggestions for you to fix ahead of time:

```
  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build date will be added to the
    package when you perform `R CMD build` on it.
  ✖ use '<-' for assignment instead of '='. '<-' is the standard, and R users and developers are used it and it is
    easier to read your code for them if you use '<-'.

    R/pclm_1D.R:117:20
    R/pclm_1D.R:141:16
    R/pclm_1D.R:141:33
    R/pclm_1D.R:141:50
    R/pclm_2D.R:69:20
    ... and 23 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80 characters

    R/pclm_fit.R:14:1
    R/pclm_fit.R:16:1
    R/pclm_graphics.R:41:1
    R/pclm_graphics.R:48:1
    R/pclm_graphics.R:72:1
    ... and 7 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use seq_len() or seq_along() instead.

    R/pclm_2D.R:137:9
    R/pclm_graphics.R:120:11
    R/pclm_graphics.R:132:11
    R/pclm_graphics.R:133:11

  ✖ fix this R CMD check NOTE: Namespace in Imports field not imported from: ‘Rdpack’ All declared Imports should be
    used.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the logicals 'TRUE' and 'FALSE' by default, but
    are not reserved words and hence can be overwritten by the user.  Hence, one should always use 'TRUE' and 'FALSE' for the
    logicals.

    R/pclm_1D.R:NA:NA
    R/pclm_1D.R:NA:NA
    R/pclm_2D.R:NA:NA
    R/pclm_2D.R:NA:NA
    R/pclm_fit.R:NA:NA
    ... and 20 more lines
```
