### 2018-09-12 964b126
adjust coding style by following karthik's suggestions
