+++ R/pclm_1D.R
@@ -114,7 +114,7 @@ pclm <- function(x, y, nlast, offset = NULL, out.step = 1, ci.level = 95,
   pclm.input.check(input, type)
   
   # Preliminary; start the clock
-  if (verbose) {pb = startpb(0, 100); on.exit(closepb(pb)); setpb(pb, 1)}
+  if (verbose) {pb <- startpb(0, 100); on.exit(closepb(pb)); setpb(pb, 1)}
   I[c("x", "y", "nlast", "offset")] <- create.artificial.bin(I) # ***
   
   # Deal with offset term
@@ -122,7 +122,7 @@ pclm <- function(x, y, nlast, offset = NULL, out.step = 1, ci.level = 95,
     if (length(offset) == length(y)) {
       if (verbose) { setpb(pb, 5); cat("   Ungrouping offset")}
       I$offset <- pclm(x = I$x, y = I$offset, I$nlast, offset = NULL, 
-                       out.step, ci.level, verbose = F, control)$fitted
+                       out.step, ci.level, verbose = FALSE, control)$fitted
     } 
   }
   
@@ -138,7 +138,7 @@ pclm <- function(x, y, nlast, offset = NULL, out.step = 1, ci.level = 95,
   R  <- delete.artificial.bin(R) # ***
   G  <- map.bins(x, nlast, out.step)
   dn <- G$output$names
-  names(R$fit) = names(R$lower) = names(R$upper) = names(R$SE) <- dn
+  names(R$fit) <- names(R$lower) <- names(R$upper) <- names(R$SE) <- dn
   
   # Output
   Fcall <- match.call()
@@ -167,7 +167,7 @@ pclm <- function(x, y, nlast, offset = NULL, out.step = 1, ci.level = 95,
 #' @export
 residuals.pclm <- function(object, ...) {
   if (!is.null(object$input$offset)) {
-    stop("residuals method not implemented for hazard rates", call. = F)
+    stop("residuals method not implemented for hazard rates", call. = FALSE)
   }
   C <- object$deep$C
   C <- C[-nrow(C), -ncol(C)]
@@ -194,6 +194,7 @@ print.pclm <- function(x, ...){
   cat("\n")
 }
 
+
 #' Summary for pclm method
 #' @inheritParams base::summary
 #' @keywords internal
@@ -209,6 +210,7 @@ summary.pclm <- function(object, ...) {
   out <- structure(class = "summary.pclm", as.list(environment()))
   return(out)
 }
+
 
 #' Print for summary.pclm method
 #' @param x An object of class \code{"summary.pclm"}

[…]
