### 2018-09-18 #937 @karthik
@rlbarter Good to check the latest version. The bump from `1.0.3` to `1.0.4` appears to be in response to my suggestions (most were cosmetic) which should not change any of your review comments. You can view those changes here: https://github.com/mpascariu/ungroup/commit/964b126b145ef4a390f3a0019f6717f30715265c
