### 2018-09-18 #937 @mpascariu
@rlbarter Thanks for your comments. See below my response.

1) In the `README.md` it is specified the version no. of the stable software available on CRAN. The development version will change more frequently and has no real importance to the common user.

2) In GitHub we provide all the source file. The `.rmd` file can be found in `ungroup/vignettes/` folder. The already compiled vignette can be found in `ungroup/inst/doc/Intro.pdf`. We will add a link in `README.md` is this solves the situation.

3) The guidelines for contributing can be found in `CONTRIBUTING.md` file.

The last point I will address in the next comment.
