### 2018-09-18 ded2aac
Add directions to vignettes and contributing files, and some indication in pclm example
