### 2018-09-18 #937 @mpascariu
@rlbarter Your suggestions have been addressed in `v1.0.5` not on GitHub. Thanks.
