### 2018-09-18 #937 @rlbarter
@mpascariu Great. This all looks good to me! Up to you whether you want to add a URL link to the vignette in the README (I personally find this helpful in general when I go to R packages on GitHub).
