### 2018-09-19 #937 @karthik
@mpascariu As @arfon noted earlier, it is very challenging to typeset accurately when using markdown + pandoc.  Given that you also have such a short paper the accurate placement you seek would be very challenging (without inserting a lot of empty space and page breaks).  You might have to mess with figure sizes to make it fit the way you seek.

I'll let @arfon weigh in but I think the paper looks fine the way it is.
