### 2018-09-19 3539e18
Minor change in joss/paper.md
