### 2018-09-19 #937 @mpascariu
@karthik I understand.

I made some minor changes in the `.md` file. I think the paper should look better now.
Can we generate the PDF again? Thank you.
