### 2018-09-19 #937 @arfon
@mpascariu - you can ask @whedon to recompile the paper for you with `@whedon generate pdf`

Also, it looks like the `#Summary` needs changing to `# Summary` at the top of the paper.
