### 2018-09-20 #937 @mpascariu
@arfon I think the article looks fine now. Thank you very much.
