### 2018-09-20 e720b4f
prepare v1.1.0 for CRAN following JOSS review
