+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: ungroup
 Type: Package
 Title: Penalized Composite Link Model for Efficient Estimation of 
  Smooth Distributions from Coarsely Binned Data
-Version: 1.0.5
+Version: 1.1.0
 Authors@R: c(
   person("Marius D.", "Pascariu", role = c("aut", "cre"), email = "rpascariu@outlook.com", comment = c(ORCID = "0000-0002-2568-6489")),
   person("Silvia", "Rizzi", role = "aut", email = "srizzi@health.sdu.dk"),
